Add upgrade for CMT delays
Add an upgrade routine for CMT delays. This fix will only work in conjunction with a fix in the CMT delay itself.

The CMT delay uses `sprintf` calls to generate the technical names and display names of the delays. These calls are locale dependent. As a consequence for example the feedback delay might have been saved either as "fbdelay_0.1s" (point) or "fbdelay_0,1s" (comma) in a save file.

The CMT fix makes sure that all delays use points in their names. Therefore we must upgrade all names that contain commas to ones that contain points.