Extract step computation in method
Extract the computation of the selected step into a dedicated method in `MidiClipView`. The removes duplicated code in the mouse press and wheel event handling code.

The method `getStep` returns an `std::optional` which contains the value of the step if the calculation was successful. It also contains the check for the height.

The calculation with floats was removed in `getStep` as it should not be necessary.

Adjust `MidiClipView::wheelEvent` so that it now computes the step in the same way as the mouse event handler. This reinstates the functionality to set the note velocities with the mouse wheel even if the beat pattern is enlarged.