Fix rendering artifacts
Fix artifacts that occur when Qt is requested to render SVGs into very small pixmaps. This is done by making sure that the SVG is always rendered into pixmaps that are at least 24x24 in size (the size of the original PNG pixmap).

Adjust the pixmap drawing code so that a potentially larger source pixmap is drawn into a smaller target pixmap.