Devirtualize instrument flags
The instruments flags (single streamed, MIDI based, not bendable) are properties of an instrument that do not change over time. Therefore the flags are made a property which is initialized at construction time.

Adjust `Instrument::flags` so that it simply returns the property. The method is not virtual anymore.

Adjust the constructors of all instruments which overrode the `flags` method to pass their flags into the `Instrument` constructor.