don't use PollParentThread for the wine process
In my setup (Ubuntu 22.04 + wine-9.10 (Staging)) the wine Vst process
doesn't have a parent and gets killed by PollParentThread shortly after
start. Strangely this only happens in 64bit mode and not in 32bit.

In either case it's a bad idea to rely on wine created process having
a parent or not because the wine project may change this anytime.