Build file modification to prepare for core and gui separation change
Up until now, all source files were added to a list and compiled
together from the toplevel CMakeLists.txt file.

The aim for this change is each folder creates a separate target,
forcing it to contain only sources contained within. A good side effect
is that the paths for sources become simpler. However, the most
important part is that this change enables further separation of the
different modules of the software.

In order to keep this small, include paths and build options have been
mostly kept as they were. The first follow up to this shall be limiting
the usage of gui from core, so that it can be later enforced by not
including the gui interfaces from the lmms_core_objs target.