Don't save MIDI information in presets
Ensure that no MIDI information is stored in presets.

Add the new protected method `isSimpleSerializing` that can be used by sub
classes to find out if they are temporarily in preset saving mode or
not. Please note that the existing implementation is very brittle
because an instance might accidentally stay in that mode.

Use the information in `InstrumentTrack::saveTrackSpecificSettings`.