Enable protected setPresetMode
Enable to make `setPresetMode` protected.

This is accomplished by adding the new methods `savePreset` and
`loadPreset` to `InstrumentTrack`. These methods manage the calling
of `setPresetMode` so that callers won't have to know about this
implementation detail. Then `saveSettings` and `loadSettings` are called
respectively which in turn will delegate to `Track`.

All context that use `loadPreset` or `savePreset` already known that
they are dealing with an `InstrumentTrack` which is the reason why this
works:
* Loading of a preset file
* Saving a preset
* Replacement of an instrument

Therefore in these place `loadPreset` and `savePreset` can be called and
it's not necessary anymore to call `setPresetMode`  there.