Add guard to safely reset bools
Add the new local class `BoolGuard` which uses RAII to safely set a
boolean to a value and later reset it to the old value.

The guard is needed to ensure that the boolean `m_presetMode` is safely
set to `true` in `savePreset` and `loadPreset` and reset even in case of
exceptions.

For this to work `m_presetMode` had to be changed to protected
visibility. On this other hand this enables the removal of
`setPresetMode`.

Add comments to the two places where previously the mode was reset so
that people do not wonder why nothing is reset and potentially put the
code back in there.