Get rid of m_presetMode and `isPresetMode()`
Get rid of the protected member `m_presetMode` and its corresponding
method `isPresetMode()`.

This is accomplished by introducing two new methods `saveTrack` and
`loadTrack`. These methods have a similar interface to `saveSettings`
and `loadSettings` but they additionally contain a boolean which
indicates if a preset is saved/loaded or a whole track. Both new
methods contain the previous code of `saveSettings` and `loadSettings`.
The latter two now only delegate to the new methods assuming that the
full track is to be stored/loaded if called via the overridden methods
`saveSettings` and `loadSettings`.

Adjust `saveTrackSpecificSettings` so that it also passes information
of whether a preset or a whole track is stored. This leads to changes
in the interfaces of `AutomationTrack`, `InstrumentTrack`,
`PatternTrack` and `SampleTrack`. Only the implementation of
`InstrumentTrack` uses the new information though.

Remove the `BoolGuard` class and adjust the implementation of
`InstrumentTrack::savePreset` and `InstrumentTrack::loadPreset` to
simply delegate to the `saveTrack` and `loadTrack` methods with the
correct parameters. The methods `savePreset` and `loadPreset` are kept
because they make the code more readable.

Implementation notes
---------------------
It could be considered to move `savePreset` and `loadPreset` into
`Track`.

The method `loadTrackSpecificSettings` was not adjusted with an
additional boolean because for now no implementation needs to
explicitly know if a preset or a track is restored. It could be
considered to make the interfaces symmetrical though.