Move preset methods into Track
Move the methods `savePreset` and `loadPreset` into track because it
potentially makes sense for all types of tracks to save a preset.

Make `Track::saveTrack` and `Track::loadTrack` private because the public
interface for saving and loading full tracks is `saveSettings` and
`loadSettings` and for saving and loading presets it's `savePreset` and
`loadPreset`.