Selected clip length updates
Do not update the clip length when dragging a node in
`AutomationClip::setDragValue`.

This is achieved by removing the call to `updateLength` from
`AutomationClip::removeNode` and explicitly calling the update in all
other places that call `removeNode` except in `setDragValue`.

This commit therefore also improves the separation of concerns.
Removing a node now only update the data structures. Updating the clip
length is a completely different thing and rather a "policy" and
therefore both calls should be done independently.