Refactor MainWindow::refocus
Refactor `MainWindow::refocus` to make it more readable and
straight-forward. Replace the Qt collection with a vector so that we
can do a simple foreach loop. Also remove the local variable `found` by
returning early on success or setting the focus on the main window
otherwise.

Remove code repetition by fetching the GUI once via `getGUI`.